requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
pyparsing==2.4.0
python-dateutil==2.8.0
python-mpd2==1.0.0
six==1.12.0
